test update with auth

+++ app/Http/Controllers/TodosController.php
@@ -25,7 +25,7 @@ class TodosController extends Controller
      * コンストラクタ
      * @return void
      */
-    public function __construct(Todo $todo)
+    public function __construct(Todo $todo, Request $request)
     {
         $this->todo = $todo;
         $this->user = \Auth::user();
@@ -88,6 +88,9 @@ class TodosController extends Controller
     public function update($id)
     {
         $todo = $this->todo->find($id);
+        if ($todo->user_id !== $this->user->id) {
+            return Redirect::route('todos.index');
+        }
 
         // 入力データを取得する
         $input = Input::only(['status']);
@@ -96,7 +99,6 @@ class TodosController extends Controller
         $rules = [
             "status" => ['required', 'numeric', 'min:1', 'max:2'],
         ];
-
         // バリデーションを実行する
         $validator = Validator::make($input, $rules);
         if ($validator->fails()) {
@@ -116,6 +118,7 @@ class TodosController extends Controller
 
         // index にリダイレクトする
         return Redirect::route('todos.index');
+
     }
 
 
@@ -123,13 +126,11 @@ class TodosController extends Controller
     {
         // Todoオブジェクトを所得する
         $todo = $this->todo->find($id);
-        // バリデーションルールの定義
+        $input = Input::only(['title']);
+
         $rules = [
             "title" => 'required|min:3|max:255',
         ];
-        // 入力データを取得する
-        $input = Input::only(['title']);
-        // バリデーションを実行する
         $validator = Validator::make($input, $rules);
         if ($validator->fails()) {
             // Ajax レスポンスを返す
@@ -143,7 +144,6 @@ class TodosController extends Controller
         $todo = $todo->fill([
             'title' => $input['title'],
         ]);
-        // データを更新する
         $todo->save();
 
         // Ajaxレスポンスを返す
@@ -156,12 +156,13 @@ class TodosController extends Controller
      */
     public function delete($id)
     {
-        // Todoオブジェクトを取得する
         $todo = $this->todo->find($id);
-        // データを削除する
-        $todo->delete();
+        if ($todo->user_id !== $this->user->id) {
+            // セッションにメッセージ 「権限ないよ！！」
+            return Redirect::route('todos.index');
+        }
 
-        // indexにリダイレクトする
+        $todo->delete();
         return Redirect::route('todos.index');
     }
